+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Added
 - Equal functions for HourlyGeothermalLoad and MonthlyGeothermalLoadAbsolute (issue #189).
+- Cylindrical borehole correction.
 
 ## changed
 - Negative reference temperatures for the fluid are now possible (issue #192).
@@ -15,6 +16,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Bug in load-duration curve when not working with optimize load profile (issue #189).
 - Bug in hourly data (issue #196).
 - Bug in saving after a file has been moved (issue #198).
+- Problem with negative g-function values (issue #187).
 - Bug in DHW and peak heating power(issue #202).